Add SCSS to override bootstrap button to match LBP theme
